[RESTEASY-1991] response builder built response (#1835)
* [RESTEASY-1991] Exception removed from BuiltResponse readEntity and code added for the task.  BuiltResponseEntityNotBacked created to support existing exception throwing scenarios.  ClientFilterResponseBuilderTest created to test code change.

* Update ClientResponse.java

I changed "    " (four spaces) to "   " (three spaces), and the diff is *much* easier to read.

ffs
