+++ resteasy-jaxrs/src/main/java/org/jboss/resteasy/core/AsyncResponseConsumer.java
@@ -23,6 +23,7 @@ import org.jboss.resteasy.plugins.providers.sse.SseConstants;
 import org.jboss.resteasy.plugins.providers.sse.SseImpl;
 import org.jboss.resteasy.resteasy_jaxrs.i18n.Messages;
 import org.jboss.resteasy.specimpl.BuiltResponse;
+import org.jboss.resteasy.specimpl.BuiltResponseEntityNotBacked;
 import org.jboss.resteasy.specimpl.MultivaluedTreeMap;
 import org.jboss.resteasy.spi.AsyncResponseProvider;
 import org.jboss.resteasy.spi.AsyncStreamProvider;
@@ -178,7 +179,8 @@ public abstract class AsyncResponseConsumer
          Response r = (Response) entity;
          Headers<Object> metadata = new Headers<Object>();
          metadata.putAll(r.getMetadata());
-         builtResponse = new BuiltResponse(r.getStatus(), r.getStatusInfo().getReasonPhrase(), metadata, r.getEntity(),  method.getMethodAnnotations());
+         builtResponse = new BuiltResponseEntityNotBacked(r.getStatus(), r.getStatusInfo().getReasonPhrase(),
+                 metadata, r.getEntity(),  method.getMethodAnnotations());
       }
       else
       {
